+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.31
 - Check `watchman` once per-session when in `Auto` mode (requires restart after installing it).
+- Prevent incorrect `Zombie initialization` error when running **Sorbetto: Stop Sorbet** but Sorbet is not installed.
 
 ## 0.3.30
 - Improve `require_relative` autocomplete: show only name in dropdown and handle `./` in any segment.